+++ NEWS.md
@@ -1,3 +1,12 @@
+
+# IEATools 0.1.53 (2021-05-07)
+
+* The "Oil and gas extraction" activity is split in "Oil extraction",
+  which extracts oil products, and "Natural gas extraction", which extracts natural gas.
+* New `split_oil_gas_extraction_eiou()` function splits the EIOU of 
+  the "Oil and gas extraction" industry into EIOU for the
+  "Oil extraction" and for the "Natural gas extraction" activities.
+
 # IEATools 0.1.52 (2021-05-07)
 
 * Deleted an efficiency for LPG stoves in Ghana in 1971